Testing out a new way to handle the timestamps on ingest
Passing datetime strings to the ingest method is causing it to run
about 500-600x slower than if we passed numeric timestamps. This is
primarily because of try/catch method we have of determining string vs
numeric. I modified the staging tables to accept a string for the
datetime and then only convert the numeric timestamps. This makes the
string method on par with the numeric method. Since we can control the
shape of the data coming to this part I would recommend that we come
up with a standard format and use the try/catch correctly, and not as
an if/then.