Feature/clarity ingest support (#11)
* Improvements to allow nodes to be added from measurements in lcs

* Moved to using orjson in the fetch process

Old method was not keeping ascii characters as is and instead of
updating the old json.dumps method we are just shifting over to orjson
like the lcs method is using which has different defaults.

* Updated check to use orjson, added line check before load

* Added sensors_latest update on ingestion

* Updated to use new split_ingest_id method to parse ingest_id

* Testing out a new way to handle the timestamps on ingest

Passing datetime strings to the ingest method is causing it to run
about 500-600x slower than if we passed numeric timestamps. This is
primarily because of try/catch method we have of determining string vs
numeric. I modified the staging tables to accept a string for the
datetime and then only convert the numeric timestamps. This makes the
string method on par with the numeric method. Since we can control the
shape of the data coming to this part I would recommend that we come
up with a standard format and use the try/catch correctly, and not as
an if/then.

* Fixed timestamp 'bug', Added diagnostic data to ingest

* Added providers and timezones check on insert

* More updates to help with ingesting

* Updated ingester

* Adding batch methods for check

* Ingest improvements

* Added start/end date lookup to realtime fetcher

* Cleaning up

* Temporary fix to the airgradient duplication issue

* Updated the lcs class to support the newer data format

  New class is meant to be more flexible and to work for all file formats.

* Cleaned up to work in production setting

* Swtiched source of loader in the handler

---------

Co-authored-by: Christian Parker <cparker@talloaks-environmental.com>